Fix some spaces and move out of ui package
